done but still need unit test

+++ service/emergency-ops/src/main/java/project/tracknest/emergencyops/configuration/messagequeue/KafkaProperties.java
@@ -10,7 +10,7 @@ import java.util.List;
 public class KafkaProperties {
     private List<String> topics;
 
-    private int partitions;
+    private short partitions;
 
     private short replicationFactor;
 }
