+++ service/emergency-ops/src/main/java/project/tracknest/emergencyops/domain/safezonelocator/service/SafeZoneLocatorService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface SafeZoneLocatorService {
     List<GetNearestSafeZonesResponse> retrieveNearestSafeZones(
-            float latitudeDegrees,
-            float longitudeDegrees,
+            double latitudeDegrees,
+            double longitudeDegrees,
             float maxDistanceMeters,
             int maxNumberOfSafeZones
     );
